Refactor 5 - Move functions which interacts with Alchemy to the alchemy folder
